pyproject.toml
[tool.poetry]
name = "pubmed-paper-fetcher"
version = "0.1.0"
description = "Fetch PubMed papers with pharma/biotech affiliations."
authors = ["Sadiya Khanam <khanamsadiya1234@gmail.com>"]
readme = "README.md"
packages = [{ include = "pubmed_paper_fetcher", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.10"
biopython = ">=1.85,<2.0"
pandas = ">=2.3.1,<3.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
get-papers-list = "pubmed_paper_fetcher.main:main"
